pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photogeoview"
version = "1.0.0"
description = "AI統合写真地理情報ビューア"
authors = [
    {name = "AI Integration Team", email = "ai-team@photogeoview.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12,<3.15"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Graphics :: Viewers",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    # GUI Framework (6.8+で全モジュール統合済み)
    "PySide6>=6.8.0",
    
    # 画像処理
    "Pillow>=10.4.0",
    
    # EXIF処理（ExifReadは開発停滞、将来的にpiexifへ移行予定）
    "ExifRead>=3.3.2",
    
    # 地図表示
    "folium>=0.18.0",
    
    # システム情報
    "psutil>=6.1.0",
    
    # カスタムコンポーネント
    "qt-theme-manager>=0.2.4",
    "breadcrumb-addressbar>=0.2.1",
]

[project.urls]
Homepage = "https://github.com/scottlz0310/PhotoGeoView"
Repository = "https://github.com/scottlz0310/PhotoGeoView.git"
Issues = "https://github.com/scottlz0310/PhotoGeoView/issues"

[dependency-groups]
dev = [
    # テスティング
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-qt>=4.4.0",
    "pytest-xdist>=3.6.0",
    "pytest-benchmark>=5.0.0",
    
    # リンティング・フォーマット
    "ruff>=0.8.0",
    
    # 型チェック
    "mypy>=1.13.0",
    
    # セキュリティ
    "bandit[toml]>=1.8.0",
    "safety>=3.2.0",
]
ci = [
    # CI用（devから必要なものを選択）
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-qt>=4.4.0",
    "pytest-xdist>=3.6.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "toml>=0.10.0",
    "tomlkit>=0.13.0",
    "watchdog>=6.0.0",
]

[project.scripts]
photogeoview = "src.main:main"
ci-simulator = "tools.ci.simulator:main"
pgv-ci = "tools.ci.simulator:main"
photogeoview-ci = "tools.ci.simulator:main"

# Ruff設定
[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
    "dev_scripts",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E203",   # whitespace before ':'
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise from
    "UP006",  # use-pep585-annotation (keep for compatibility)
    "UP007",  # use-pep604-annotation (keep for compatibility)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "SIM"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
known-third-party = ["PySide6", "PIL", "folium", "pytest"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# mypy設定
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "folium.*",
    "piexif.*",
    "psutil.*",
]
ignore_missing_imports = true

# pytest設定
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "ai_compatibility: marks tests as AI compatibility tests",
    "performance: marks tests as performance tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# Coverage設定
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/build/*",
    "*/dist/*",
    "*/dev_scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit設定（セキュリティ）
[tool.bandit]
exclude_dirs = ["tests", "venv", "build", "dist", "dev_scripts"]
skips = ["B101", "B601"]  # assert文とshell=Trueの警告をスキップ

# AI統合品質設定
[tool.ai_integration]
quality_threshold = 70.0
coverage_threshold = 80.0
performance_regression_threshold = 30.0

[tool.ai_integration.components]
copilot = {focus = "core_functionality", quality_weight = 1.2}
cursor = {focus = "ui_ux", quality_weight = 1.0}
kiro = {focus = "integration", quality_weight = 1.5}

# CI Simulator設定
[tool.ci_simulator]
enabled = true
default_python_versions = ["3.9", "3.10", "3.11"]
timeout = 1800  # 30 minutes
parallel_jobs = 4
fail_fast = false

[tool.ci_simulator.checks]
code_quality = {enabled = true, auto_fix = false}
test_runner = {enabled = true, coverage_threshold = 80.0}
security_scanner = {enabled = true, fail_on_high = false}
performance_analyzer = {enabled = true, regression_threshold = 30.0}
ai_component_tester = {enabled = true, demo_tests = true}

[tool.ci_simulator.directories]
reports = "reports/ci-simulation"
logs = "logs"
history = ".kiro/ci-history"
temp = "temp/ci-simulation"

[tool.ci_simulator.git_hooks]
pre_commit = {enabled = true, checks = ["code_quality", "test_runner"]}
pre_push = {enabled = false, checks = ["all"]}

[tool.ci_simulator.notifications]
slack_webhook = ""
email_recipients = []
github_status = true
